updates-monitor: Added: Multithreaded download support for ordinary files

+++ updates-monitor/src/main/java/core/youtube/YoutubeVideoHandler.java
@@ -1,14 +1,8 @@
 package core.youtube;
 
 import java.io.File;
-import java.io.FileOutputStream;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
-import java.net.HttpURLConnection;
-import java.net.URL;
 import java.util.ArrayList;
-import java.util.Iterator;
 import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -23,17 +17,15 @@ import util.DownloadUtils;
 import util.FFMPEGUtils;
 import util.PageParsing;
 import utils.FileUtils;
-import utils.FormattingUtils;
 import utils.JSONUtils;
 
 public class YoutubeVideoHandler {
     private static final Pattern VIDEO_URL_PATTERN = Pattern.compile("^https?:\\/\\/www.youtube.com\\/watch\\?v=([0-9A-Za-z_-]+)&?.*$");
-    private static final int UPDATE_PROGRESS_INTERVAL = 100;
 
-    private boolean printProgress;
+    private YoutubeOrdinaryFileDownloader youtubeOrdinaryFileDownloader;
 
-    public void setPrintProgress(boolean printProgress) {
-        this.printProgress = printProgress;
+    public YoutubeVideoHandler() {
+        youtubeOrdinaryFileDownloader = new YoutubeOrdinaryFileDownloader();
     }
 
     public String parseVideoId(String url) {
@@ -46,8 +38,8 @@ public class YoutubeVideoHandler {
         }
     }
 
-    public Result downloadVideo(YoutubeVideo video, String downloadChannelPath, String temporaryFolderPath) throws IOException {
-        Result result = new Result();
+    public YoutubeResult downloadVideo(YoutubeVideo video, String downloadChannelPath, String temporaryFolderPath) throws Exception {
+        YoutubeResult result = new YoutubeResult();
 
         // Downloading Video Page
         String urlString = "https://www.youtube.com/watch?v=" + video.getVideoId();
@@ -81,53 +73,83 @@ public class YoutubeVideoHandler {
         List<VideoFormat> videoFormats = fetchVideoFormats(streamingDataNode);
         List<AudioFormat> audioFormats = fetchAudioFormats(streamingDataNode);
 
+        if (videoFormats.size() < 1 || audioFormats.size() < 1) {
+            System.out.println("Unsupported video format: " + video.getVideoId());
+            result.unsupported = true;
+            return result;
+        }
+
         // Downloading Files itself
         String videoFileDownloadPrefix = temporaryFolderPath + "/" + video.getVideoId() + "_video";
         String audioFileDownloadPrefix = temporaryFolderPath + "/" + video.getVideoId() + "_audio";
         String temporaryFilePath = temporaryFolderPath + "/" + video.getVideoId();
 
         // Downloading Video file
-        if (printProgress) {
-            System.out.print("\n\tDownloading video... ");
+        System.out.print("\n\tDownloading video... ");
+
+        YoutubeResult downloadVideoResult = null;
+        VideoFormat videoFormat = videoFormats.get(0);
+        switch (videoFormat.type) {
+        case OrdinaryFile:
+            downloadVideoResult = youtubeOrdinaryFileDownloader.download(video.getVideoId(), videoFileDownloadPrefix, temporaryFilePath, videoFormat);
+            break;
+        case OTF_Stream:
+            // TODO Implement
+            break;
+        case Encrypted:
+            // TODO Implement
+            break;
+        default:
+            System.out.println("Unsupported video format: " + video.getVideoId() + ", format type: " + videoFormat.type.name());
+            result.unsupported = true;
+            return result;
         }
-        Result downloadVideoResult = downloadFile(videoFileDownloadPrefix, temporaryFilePath, videoFormats);
+        if (downloadVideoResult == null) {
+            System.out.println("Unsupported video format: " + video.getVideoId());
+            result.unsupported = true;
+            return result;
+        }
         if (downloadVideoResult.notFound) {
-            if (printProgress) {
-                System.out.print("Not found...");
-            }
+            System.out.print("Not found...");
             result.notFound = true;
             return result;
         }
         result.completed = downloadVideoResult.completed;
 
         // Downloading Audio file
-        if (printProgress) {
-            System.out.print("\n\tDownloading audio... ");
+        if (!result.completed) {
+            return result;
         }
-        Result downloadAudioResult = downloadFile(audioFileDownloadPrefix, temporaryFilePath, audioFormats);
+        System.out.print("\n\tDownloading audio... ");
+        AudioFormat audioFormat = audioFormats.get(0);
+        YoutubeResult downloadAudioResult = youtubeOrdinaryFileDownloader.download(video.getVideoId(), audioFileDownloadPrefix, temporaryFilePath, audioFormat);
         if (downloadAudioResult.notFound) {
-            if (printProgress) {
-                System.out.print("Not found...");
-            }
+            System.out.print("Not found...");
             result.notFound = true;
             return result;
         }
         result.completed &= downloadAudioResult.completed;
 
         // Combine Video and Audio via ffmpeg
-        if (printProgress) {
-            System.out.print("\n\tCombining video and audio tracks via ffmpeg... ");
+        if (!result.completed) {
+            return result;
         }
+        System.out.print("\n\tCombining video and audio tracks via ffmpeg... ");
         String videoTrackPath = downloadVideoResult.resultFile.getAbsolutePath();
         String audioTrackPath = downloadAudioResult.resultFile.getAbsolutePath();
         String ffmpegResultPath = temporaryFolderPath + "/" + video.getVideoId() + "_combined.mp4";
         result.completed &= FFMPEGUtils.combineVideoAndAudio(videoTrackPath, audioTrackPath, ffmpegResultPath);
-        if (printProgress) {
-            System.out.print(result.completed ? "Done" : "Failed");
-        }
-        result.completed &= FileUtils.moveFile(new File(ffmpegResultPath), new File(downloadChannelPath, video.getFilename()));
         result.completed &= downloadVideoResult.resultFile.delete();
         result.completed &= downloadAudioResult.resultFile.delete();
+        System.out.print(result.completed ? "Done" : "Failed");
+
+        // Moving Temp file to Destination folder
+        if (!result.completed) {
+            return result;
+        }
+        System.out.print("\n\tMoving temporary file to the channel folder... ");
+        result.completed &= FileUtils.moveFile(new File(ffmpegResultPath), new File(downloadChannelPath, video.getFilename()));
+        System.out.print(result.completed ? "Done" : "Failed");
 
         return result;
     }
@@ -138,13 +160,16 @@ public class YoutubeVideoHandler {
         List<VideoFormat> videoFormats = new ArrayList<>();
 
         for (JsonNode formatNode : adaptiveFormatsNode) {
-            // Strange, but happens sometimes
-            if (!formatNode.has("url")) {
-                continue;
-            }
-
             if (formatNode.get("mimeType").asText().startsWith("video")) {
                 VideoFormat format = new VideoFormat();
+
+                if (!formatNode.has("url")) {
+                    format.type = VideoFormat.Types.Encrypted;
+                } else if (formatNode.has("type") && "FORMAT_STREAM_TYPE_OTF".equals(formatNode.get("type").asText())) {
+                    format.type = VideoFormat.Types.OTF_Stream;
+                } else {
+                    format.type = VideoFormat.Types.OrdinaryFile;
+                }
 
                 format.mimeType = formatNode.get("mimeType").asText();
                 if (format.mimeType.startsWith("video/mp4")) {
@@ -152,12 +177,16 @@ public class YoutubeVideoHandler {
                 } else if (format.mimeType.startsWith("video/webm")) {
                     format.fileExtension = "webm";
                 } else {
-                    // Unsupported video format
-                    continue;
+                    throw new IllegalArgumentException("Unsupported MimeType: " + format.mimeType);
                 }
 
                 format.iTag = formatNode.get("itag").asInt();
-                format.downloadURL = formatNode.get("url").asText();
+                if (formatNode.has("url")) {
+                    format.downloadURL = formatNode.get("url").asText();
+                }
+                if (formatNode.has("contentLength")) {
+                    format.contentLength = formatNode.get("contentLength").asLong();
+                }
                 format.bitrate = formatNode.get("bitrate").asInt();
                 if (formatNode.has("width")) {
                     format.width = formatNode.get("width").asInt();
@@ -221,6 +250,7 @@ public class YoutubeVideoHandler {
 
                 format.iTag = formatNode.get("itag").asInt();
                 format.downloadURL = formatNode.get("url").asText();
+                format.contentLength = formatNode.get("contentLength").asLong();
                 format.bitrate = formatNode.get("bitrate").asInt();
                 format.sampleRate = formatNode.get("audioSampleRate").asInt();
                 format.channels = formatNode.get("audioChannels").asInt();
@@ -249,146 +279,22 @@ public class YoutubeVideoHandler {
         return audioFormats;
     }
 
-    private Result downloadFile(String downloadFilePrefix, String temporaryFilePath, List<? extends Format> formats) throws IOException {
-        Result result = new Result();
-
-        File tempFile = new File(temporaryFilePath);
-        // TODO Download resume support
-
-        Iterator<? extends Format> formatIterator = formats.iterator();
-        Format format = null;
-        String videoFilename;
-        File resultFile = null;
-        HttpURLConnection connection = null;
-        int responseCode = 404;
-        while (responseCode == 404 && formatIterator.hasNext()) {
-            format = formatIterator.next();
-
-            videoFilename = downloadFilePrefix + "." + format.fileExtension;
-            resultFile = new File(videoFilename);
-            result.resultFile = resultFile;
-            if (resultFile.exists()) {
-                resultFile.delete();
-            }
-
-            URL url = new URL(format.downloadURL);
-            connection = (HttpURLConnection) url.openConnection();
-            connection.setConnectTimeout(10 * 1000);
-            connection.setReadTimeout(60 * 1000);
-            responseCode = connection.getResponseCode();
-            if (connection.getResponseCode() == 404) {
-                format = formatIterator.next();
-            }
-        }
-
-        if (responseCode == 404) {
-            result.notFound = true;
-            return result;
-        }
-
-        long contentLength = connection.getContentLengthLong();
-        if (printProgress) {
-            System.out.print(format + " ");
-            System.out.print(connection.getResponseCode() + " " + connection.getResponseMessage() + "... ");
-            System.out.print("Content-Length: " + contentLength + "... ");
-        }
-        byte[] buffer = new byte[65536];
-        try (InputStream is = connection.getInputStream();
-                OutputStream os = new FileOutputStream(tempFile)) {
-
-            long fileSize = contentLength;
-
-            long fileDownloadStartTime = System.currentTimeMillis();
-            long lastStatusUpdate = fileDownloadStartTime;
-            long readTotal = 0;
-            int lastMessageLength = 0;
-
-            int read = 0;
-            while (read != -1) {
-                read = is.read(buffer);
-                if (read > 0) {
-                    os.write(buffer, 0, read);
-                    readTotal += read;
-                }
-
-                if (printProgress) {
-                    long now = System.currentTimeMillis();
-                    if (now - lastStatusUpdate >= UPDATE_PROGRESS_INTERVAL) {
-                        long periodLength = now - fileDownloadStartTime;
-                        // Calculating download speed
-                        long speed = readTotal * 1000 / periodLength;
-
-                        // Formatting progress message
-                        StringBuilder message = new StringBuilder();
-                        message.append(FormattingUtils.humanReadableSize(readTotal))
-                                .append("b of ")
-                                .append(FormattingUtils.humanReadableSize(fileSize) + "b (");
-                        message.append(readTotal * 100 / fileSize).append("%), ");
-                        message.append(FormattingUtils.humanReadableSize(speed) + "b/s");
-                        while (message.length() < lastMessageLength) {
-                            message.append(" ");
-                        }
-
-                        // Moving cursor at the beginning of the previous progress message
-                        for (int i = 0; i < lastMessageLength; i++) {
-                            System.out.print("\b");
-                        }
-
-                        // Printing the progress message
-                        System.out.print(message);
-
-                        lastMessageLength = message.length();
-                        lastStatusUpdate = now;
-                    }
-                }
-            }
-
-            if (printProgress) {
-                // Moving cursor at the beginning of the previous progress message
-                for (int i = 0; i < lastMessageLength; i++) {
-                    System.out.print("\b");
-                }
-
-                // Final progress report on download end
-                long now = System.currentTimeMillis();
-                long fileDownloadTime = now - fileDownloadStartTime;
-                long speed = fileSize * 1000 / fileDownloadTime;
-                StringBuilder message = new StringBuilder();
-                message.append(FormattingUtils.humanReadableSize(fileSize))
-                        .append("b in ")
-                        .append(fileDownloadTime / 1000).append(" s, ");
-                message.append("average speed: ")
-                        .append(FormattingUtils.humanReadableSize(speed)).append("b/s");
-                while (message.length() < lastMessageLength) {
-                    message.append(" ");
-                }
-                System.out.print(message);
-            }
-        }
-
-        if (tempFile.length() != contentLength) {
-            throw new IOException("Incorrect file size after download, file: \"" + tempFile.getAbsolutePath()
-                + "\", expected: " + contentLength + ", actual: " + tempFile.length());
-        }
-
-        result.completed = FileUtils.moveFile(tempFile, resultFile);
-
-        return result;
-    }
 
     public static class Result {
+        public boolean unsupported;
         public File resultFile;
         public boolean completed;
         public boolean notFound;
     }
 
-    private static class Format {
+    public static class Format {
         @SuppressWarnings("unused")
         protected int iTag;
 
         protected int bitrate;
 
         protected String downloadURL;
+        protected long contentLength;
         protected String mimeType;
         protected String fileExtension;
 
@@ -398,18 +304,23 @@ public class YoutubeVideoHandler {
         }
     }
 
-    private static class VideoFormat extends Format {
+    public static class VideoFormat extends Format {
+        private Types type;
         private int width;
         private int height;
         private int fps;
 
         @Override
         public String toString() {
-            return "[" + fileExtension + ", " + width + "x" + height + "@" + fps + ", " + (bitrate / 1024) + " kbps]";
+            return "[" + fileExtension + ", " + width + "x" + height + "@" + fps + ", " + (bitrate / 1024) + " kbps, " + type.name() + "]";
+        }
+
+        public static enum Types {
+            OrdinaryFile, OTF_Stream, Encrypted
         }
     }
 
-    private static class AudioFormat extends Format {
+    public static class AudioFormat extends Format {
         private int sampleRate;
         private int channels;
